pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["boytsovau <alexbonch@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]

classifiers = [
    "Topic :: File Formats"
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]

gendiff = "gendiff.scripts.diff:main"
